EES-5400: Add/update tests.
